requirements.txt
asgiref==3.8.1
azure-core==1.34.0
azure-functions==1.23.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
cryptography==44.0.3
Django==5.1.7
django-crispy-forms==2.3
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.16.0
git-filter-repo==2.47.0
idna==3.10
isodate==0.7.2
mysqlclient==2.2.7
packaging==25.0
pillow==11.1.0
pycparser==2.22
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
